requirements.txt
torch
torchvision
timm
pandas
numpy
pillow
pyyaml
scikit-learn
matplotlib
tqdm
seaborn
pytest
